# docs/api.yaml
openapi: 3.0.3
info:
  title: AI Therapist API
  description: |
    A comprehensive API for a therapeutic AI chat application that provides compassionate mental health support through AI conversation.

    ## Features
    - Real-time streaming chat responses
    - Session management for conversation continuity
    - Professional report generation and email delivery
    - Comprehensive validation and error handling
    - Mobile-optimized responsive design

    ## Security
    - CORS protection
    - Input validation with Zod schemas
    - Rate limiting considerations
    - Secure API key handling

    ## Authentication
    The application authenticates requests using a cookie-backed session bound to a trusted device
    (TOTP-based verification with device fingerprinting). Client requests include the session cookie;
    server routes validate it and enforce rate limits. Provider API keys (e.g., Groq) are used only
    server-side and are never exposed to clients.
  version: 1.0.0
  contact:
    name: AI Therapist Support
    email: support@ai-therapist.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-domain.com/api
    description: Production server
paths:
...
  /reports/memory:
    get:
      summary: Retrieve memory context reports
      description: Returns recent reports for cross-session therapeutic memory.
      tags:
        - Reports
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
        - name: excludeSessionId
          in: query
          required: false
          schema:
            type: string
        - name: manage
          in: query
          required: false
          schema:
            type: boolean
        - name: includeFullContent
          in: query
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: Memory context response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/MemoryData'
                      - $ref: '#/components/schemas/MemoryManageData'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete memory context reports
      description: Deletes reports from memory context with filtering options.
      tags:
        - Reports
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: excludeSessionId
          in: query
          required: false
          schema:
            type: string
        - name: sessionIds
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: Memory deletion response
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DeleteResponseData'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
  headers:
    X-Request-Id:
      description: Correlation ID for tracing this request
      schema:
        type: string
...
  schemas:
...
    MemoryContextEntry:
      type: object
      required:
        - sessionTitle
        - sessionDate
        - reportDate
        - content
        - summary
      properties:
        sessionTitle:
          type: string
        sessionDate:
          type: string
          format: date
        reportDate:
          type: string
          format: date
        content:
          type: string
        summary:
          type: string
    MemoryStats:
      type: object
      required:
        - totalReportsFound
        - successfullyDecrypted
        - failedDecryptions
      properties:
        totalReportsFound:
          type: integer
        successfullyDecrypted:
          type: integer
        failedDecryptions:
          type: integer
    MemoryData:
      type: object
      required:
        - memoryContext
        - reportCount
        - stats
      properties:
        memoryContext:
          type: array
          items:
            $ref: '#/components/schemas/MemoryContextEntry'
        reportCount:
          type: integer
        stats:
          $ref: '#/components/schemas/MemoryStats'
    MemoryReportDetail:
      type: object
      required:
        - id
        - sessionId
        - sessionTitle
        - sessionDate
        - reportDate
        - contentPreview
        - keyInsights
        - hasEncryptedContent
        - reportSize
      properties:
        id:
          type: string
        sessionId:
          type: string
        sessionTitle:
          type: string
        sessionDate:
          type: string
          format: date
        reportDate:
          type: string
          format: date
        contentPreview:
          type: string
        keyInsights:
          type: array
          items:
            type: string
        hasEncryptedContent:
          type: boolean
        reportSize:
          type: integer
        fullContent:
          type: string
        structuredCBTData:
          type: object
          additionalProperties: true
    MemoryManagementStats:
      type: object
      required:
        - totalReportsFound
        - successfullyProcessed
        - failedDecryptions
        - hasMemory
      properties:
        totalReportsFound:
          type: integer
        successfullyProcessed:
          type: integer
        failedDecryptions:
          type: integer
        hasMemory:
          type: boolean
    MemoryManageData:
      type: object
      required:
        - memoryDetails
        - reportCount
        - stats
      properties:
        memoryDetails:
          type: array
          items:
            $ref: '#/components/schemas/MemoryReportDetail'
        reportCount:
          type: integer
        stats:
          $ref: '#/components/schemas/MemoryManagementStats'
    DeleteResponseData:
      type: object
      required:
        - deletedCount
        - message
        - deletionType
      properties:
        deletedCount:
          type: integer
        message:
          type: string
        deletionType:
          type: string
          enum: [specific, recent, all-except-current, all]
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: 'Validation failed'
        details:
          type: string
          description: Detailed error information
          example: 'Invalid email address format'
        code:
          type: string
          description: Error code for programmatic handling
          example: 'VALIDATION_ERROR'
        suggestedAction:
          type: string
          description: Suggested action to resolve the error
          example: 'Please check your input and try again'
  responses:
...
    InternalServerError:
      description: Internal server error
      headers:
        X-Request-Id:
          description: Correlation ID for tracing this request
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            server_error:
              summary: Generic server error
              value:
                error: 'Internal server error'
                details: 'An unexpected error occurred'
                code: 'INTERNAL_ERROR'
            database_error:
              summary: Database error
              value:
                error: 'Failed to create session'
                details: 'Unable to create a new therapy session. Please try again.'
                code: 'SESSION_CREATE_ERROR'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (currently only used for Groq API)
...